requirements-dev.txt
pydantic
pyyaml
pytest
requests
